requirements.txt
python-dotenv
kagglehub
torch
pandas
gdown
dvc
mlflow >= 3.1
notebook
numpy
matplotlib
seaborn
python-box == 6.0.2
pyYAML
tqdm
ensure == 1.0.2
joblib
types-PyYAML
scipy
fastapi
uvicorn
pydantic
-e .